feat: add fakeddit dataset to dvc
