+++ notes.md
@@ -2,3 +2,5 @@ QCM: the model takes the concatenation of tokens of the question text (Q), the c
 QCM → A: No-CoT which predicts the answer directly
 QCM → RA: Reasoning where answer inference is conditioned to the rationale 
 QCM → AR: Explanation where the rationale is used for explaining the answer inference 
+
+Dataset Fake News https://fakeddit.netlify.app/
