+++ src/data/science_qa_dataset_img.py
@@ -1,11 +1,7 @@
-import json
-from abc import ABC, abstractmethod
-
 import numpy as np
 import torch
-from torch.utils.data import Dataset
 
-from src.models.prompt import build_train_pair
+from src.data.science_qa_dataset_std import ScienceQADatasetStd
 
 # TODO img_shape should not be here!
 img_shape = {
@@ -17,13 +13,7 @@ img_shape = {
 device = 'cuda' if torch.cuda.is_available() else 'cpu'
 
 
-class ScienceQADataset(Dataset, ABC):
-
-    @abstractmethod
-    def __len__(self): pass
-
-
-class ScienceQADatasetImg(ScienceQADataset):
+class ScienceQADatasetImg(ScienceQADatasetStd):
     """
     Creating a custom dataset for reading the dataset and
     loading it into the dataloader to pass it to the
@@ -35,13 +25,13 @@ class ScienceQADatasetImg(ScienceQADataset):
             self,
             problems,
             qids,
-            name_maps,
             tokenizer,
             source_len,
             target_len,
             args,
-            image_features,
-            test_le=None
+            test_le=None,
+            image_features=None,
+            name_maps=None,
     ):
         """
         Initializes a Dataset class
@@ -55,39 +45,18 @@ class ScienceQADatasetImg(ScienceQADataset):
             target_text (str): column name of target text
         """
 
-        self.tokenizer = tokenizer
-        self.data = {qid: problems[qid] for qid in qids}
-        self.source_len = source_len
-        self.summ_len = target_len
+        super(ScienceQADatasetImg, self).__init__(
+            problems,
+            qids,
+            tokenizer,
+            source_len,
+            target_len,
+            args,
+            test_le
+        )
 
-        self.source_ids = torch.tensor([]).to(device)
-        self.source_masks = torch.tensor([]).to(device)
-        self.target_ids = torch.tensor([]).to(device)
         self.image_ids = torch.tensor([]).to(device)
-
-        if test_le is not None:
-            test_le_data = json.load(open(test_le))["preds"]
-        else:
-            test_le_data = None
-
-        idx = 0
         for qid in self.data:
-            if test_le_data is not None:
-                curr_le_data = test_le_data[idx]
-                idx += 1
-            else:
-                curr_le_data = None
-            prompt, target = build_train_pair(
-                problems, qid, args, curr_le_data)
-
-            # SOURCE
-            source = self.process_data(prompt, self.source_len)
-            source_id = source["input_ids"].squeeze().to(device)
-            source_mask = source["attention_mask"].squeeze().to(device)
-
-            # TARGET
-            target = self.process_data(target, self.summ_len)
-            target = target["input_ids"].squeeze().to(device)
 
             if str(qid) in name_maps:
                 i_vectors = image_features[int(name_maps[str(qid)])]
@@ -96,18 +65,8 @@ class ScienceQADatasetImg(ScienceQADataset):
                 i_vectors = np.zeros(shape)
             i_vectors = torch.tensor(i_vectors).squeeze().to(device)
 
-            self.source_ids = torch.cat(
-                (self.source_ids, source_id.unsqueeze(0)), 0)
-            self.source_masks = torch.cat(
-                (self.source_masks, source_mask.unsqueeze(0)), 0)
-            self.target_ids = torch.cat(
-                (self.target_ids, target.unsqueeze(0)), 0)
             self.image_ids = torch.cat(
                 (self.image_ids, i_vectors.unsqueeze(0)), 0)
-
-    def __len__(self):
-        """returns the length of dataframe"""
-        return len(self.target_ids)
 
     def __getitem__(self, index):
         """return the input ids, attention masks and target ids"""
@@ -119,45 +78,4 @@ class ScienceQADatasetImg(ScienceQADataset):
             "labels": self.target_ids[index].to(torch.long).tolist(),
         }
 
-    def process_data(
-            self,
-            text,
-            max_length
-    ):
-        text = " ".join(str(text).split())
-        return self.tokenizer.batch_encode_plus(
-            [text],
-            max_length=max_length,
-            pad_to_max_length=True,
-            truncation=True,
-            padding="max_length",
-            return_tensors="pt",
-        )
 
-
-class ScienceQADatasetIterator:
-
-    def __init__(self, dataset: ScienceQADataset, batch_size: int = 100):
-        self._dataset = dataset
-        self.batch_size = batch_size if batch_size else len(self._dataset)
-        self.num_batches = int(len(self._dataset) / self.batch_size)
-        if len(self._dataset) %  self.batch_size:
-            self.num_batches += 1
-
-    def __iter__(self):
-        self._index = 0
-        return self
-
-    def __next__(self):
-        if self._index < self.num_batches:
-            items = []
-            for i in range(self.batch_size):
-                try:
-                    index = (self._index * self.batch_size) + i
-                    items.append(self._dataset.__getitem__(index))
-                except IndexError:
-                    break
-            self._index += 1
-            return items
-        else:
-            raise StopIteration
